requirements.txt
tableauserverclient>=0.19.0
pathlib>=1.0.1 
requests>=2.25.1
PyGithub>=1.54.1
python-dotenv>=1.0.0 